* make project compile * add create recipe endpoint * change ID types to long * draft service layer

+++ src/main/kotlin/de/shevchuk/buy_recipe/dto/ShoppingCart.kt
@@ -3,6 +3,6 @@ package de.shevchuk.buy_recipe.dto
 
 data class ShoppingCart(
     val id: Long,
-    val totalInCents: Long,
+    val totalInCents: Int,
     val items: List<CartItem> = emptyList()
 )
